fix(dgraph): use core isDate fns
